requirements.txt
amqp==5.0.6
asgiref==3.4.1
attrs==21.2.0
billiard==3.6.4.0
cached-property==1.5.2
cachetools==4.2.4
celery==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.1.1
cryptography==35.0.0
defusedxml==0.7.1
Django==3.2
django-allauth==0.45.0
django-celery-results==2.2.0
django-cors-headers==3.10.0
django-crispy-forms==1.13.0
django-environ==0.8.1
djangorestframework==3.12.4
factory-boy==3.2.1
Faker==9.8.0
flake8==4.0.1
google-api-core==2.2.2
google-api-python-client==2.30.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.53.0
gunicorn==20.1.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.2.0
iniconfig==1.1.1
install==1.3.4
kombu==5.2.1
mccabe==0.6.1
memory-profiler==0.58.0
oauthlib==3.1.1
packaging==21.2
pluggy==1.0.0
prompt-toolkit==3.0.22
protobuf==3.19.1
psutil==5.8.0
psycopg2==2.9.1
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.20
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==2.4.7
pytest==6.2.5
pytest-cov==3.0.0
pytest-django==4.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.2
typing-extensions==3.10.0.2
uritemplate==4.1.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
zipp==3.6.0
